refactor: split ios talk permissions
